WIP: Added naive bspmv cpu kernel implementation
This commit adds wrapper code as well as a naive implementation
for a blocked crs spmv cpu kernel.

The kernel is yet untested, will do this ASAP.

Furthmore there are a few questions regarding the interface and
the implementation. Should BCRS be a subtype of SparseMatrix
or its own type. Should padding and block major sorting be
a flag or a member in the internal data structure?

TODO: Add naive CUDA kernel and test performance.

Signed-off-by: Maximilian Esser <otk.esser@gmail.com>